moved retract and local gaussian to manifold_gaussian to avoid circular imports
